64 bit key and 16 bit val

+++ src/types.h
@@ -180,7 +180,7 @@ enum Bound : uint8_t {
     BOUND_EXACT = BOUND_UPPER | BOUND_LOWER
 };
 
-enum Value : int8_t {
+enum Value : int16_t {
     VALUE_ZERO = 0,
     VALUE_DRAW = 0,
